+++ src/ngraph/frontend/onnx_import/utils/common.hpp
@@ -113,7 +113,7 @@ namespace ngraph
             ///
             std::vector<std::size_t> validate_axes(const ngraph::onnx_import::Node& node,
                                                    std::vector<std::int64_t> axes,
-                                                   std::size_t tensor_rank);
+                                                   std::int64_t tensor_rank);
 
             /// \brief Creates a shifted square identity matrix.
             /// \note Shifting in the context of this operator means that
